+++ CoreComponents/Core/Attributes/DescriptionAttribute.cs
@@ -1,0 +1,52 @@
+﻿/*
+DescriptionAttribute.cs
+
+This file is part of Morgan's CLR Advanced Runtime (MCART)
+
+Author(s):
+     César Andrés Morgan <xds_xps_ivx@hotmail.com>
+
+Copyright (c) 2011 - 2018 César Andrés Morgan
+
+Morgan's CLR Advanced Runtime (MCART) is free software: you can redistribute it
+and/or modify it under the terms of the GNU General Public License as published
+by the Free Software Foundation, either version 3 of the License, or (at your
+option) any later version.
+
+Morgan's CLR Advanced Runtime (MCART) is distributed in the hope that it will
+be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General
+Public License for more details.
+
+You should have received a copy of the GNU General Public License along with
+this program. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+// ReSharper disable UnusedMember.Global
+// ReSharper disable ClassNeverInstantiated.Global
+// ReSharper disable MemberCanBePrivate.Global
+
+using System;
+
+namespace TheXDS.MCART.Attributes
+{
+    /// <inheritdoc />
+    /// <summary>
+    ///     Establece una descripción larga para este elemento.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.All)]
+    [Serializable]
+    public sealed class DescriptionAttribute : TextAttribute
+    {
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase
+        ///     <see cref="DescriptionAttribute" />.
+        /// </summary>
+        /// <param name="description">Valor del atributo.</param>
+        public DescriptionAttribute(string description) : base(description)
+        {
+        }
+    }
+}
