+++ CoreComponents/Core/Attributes/FloatAttribute.cs
@@ -1,0 +1,57 @@
+﻿/*
+FloatAttribute.cs
+
+This file is part of Morgan's CLR Advanced Runtime (MCART)
+
+Author(s):
+     César Andrés Morgan <xds_xps_ivx@hotmail.com>
+
+Copyright (c) 2011 - 2018 César Andrés Morgan
+
+Morgan's CLR Advanced Runtime (MCART) is free software: you can redistribute it
+and/or modify it under the terms of the GNU General Public License as published
+by the Free Software Foundation, either version 3 of the License, or (at your
+option) any later version.
+
+Morgan's CLR Advanced Runtime (MCART) is distributed in the hope that it will
+be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General
+Public License for more details.
+
+You should have received a copy of the GNU General Public License along with
+this program. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+// ReSharper disable UnusedMember.Global
+// ReSharper disable ClassNeverInstantiated.Global
+// ReSharper disable MemberCanBePrivate.Global
+
+using System;
+
+namespace TheXDS.MCART.Attributes
+{
+    /// <inheritdoc cref="Attribute"/>
+    /// <summary>
+    ///     Clase base para los atributos basados en números flotantes.
+    /// </summary>
+    public abstract class FloatAttribute : Attribute, IValueAttribute<float>
+    {
+        /// <inheritdoc />
+        /// <summary>
+        ///     Crea una nueva isntancia de la clase
+        ///     <see cref="FloatAttribute" />.
+        /// </summary>
+        /// <param name="attributeValue">Valor de este atributo.</param>
+        protected FloatAttribute(float attributeValue)
+        {
+            Value = attributeValue;
+        }
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Obtiene el valor asociado a este atributo.
+        /// </summary>
+        public float Value { get; }
+    }
+}
