+++ CoreComponents/Core/Attributes/UnstableAttribute.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace TheXDS.MCART.Attributes
+{
+    /// <inheritdoc />
+    /// <summary>
+    ///     Indica que un elemento contiene código que podría ser inestable.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class | AttributeTargets.Module | AttributeTargets.Assembly)]
+    [Serializable]
+    public sealed class UnstableAttribute : Attribute
+    {
+    }
+}
