+++ CoreComponents/Core/Attributes/VersionAttribute.cs
@@ -1,0 +1,37 @@
+﻿using System;
+
+namespace TheXDS.MCART.Attributes
+{
+    /// <inheritdoc />
+    /// <summary>
+    ///     Especifica la versión del elemento.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class | AttributeTargets.Module | AttributeTargets.Assembly)]
+    [Serializable]
+    public sealed class VersionAttribute : VersionAttributeBase
+    {
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase
+        ///     <see cref="VersionAttribute" />.
+        /// </summary>
+        /// <param name="major">Número de versión mayor.</param>
+        /// <param name="minor">Número de versión menor.</param>
+        public VersionAttribute(int major, int minor) : base(major, minor, int.MaxValue, int.MaxValue)
+        {
+        }
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Inicializa una nueva instancia de la clase
+        ///     <see cref="VersionAttribute" />.
+        /// </summary>
+        /// <param name="major">Número de versión mayor.</param>
+        /// <param name="minor">Número de versión menor.</param>
+        /// <param name="build">Número de compilación.</param>
+        /// <param name="rev">Número de revisión.</param>
+        public VersionAttribute(int major, int minor, int build, int rev) : base(major, minor, build, rev)
+        {
+        }
+    }
+}
