+++ CoreComponents/Core/Exceptions/InvalidTypeException.cs
